Re-use Mill daemon in example/integration tests run in `.local` mode (#5548)
Cuts the time taken for `./mill -ik 'example.javalib.__.local.daemon'`
on my laptop from ~260s to ~150s, with similar speedups for many of the
other applicable test suites.

This change reduces the hermeticity of the tests to which it is applied,
in exchange for performance. Some code changes were necessary in order
to make the Mill daemon stable enough for this to work, but presumably
those were pre-existing bugs that should be fixed anyway.

A lot of the `integration` tests and `example.migrating` tests are not
amenable to re-using the Mill daemon for a variety of reasons, so those
are left as `.packaged` and `.native`. These typically exercise the edge
cases in Mill's daemon and lifecycle management, whereas the example
tests on `.local` mode tend to be exercising features that are unrelated
to the method of executing Mill

+++ testkit/src/mill/testkit/IntegrationTesterBase.scala
@@ -24,7 +24,8 @@ trait IntegrationTesterBase {
    * non-deterministic interference and flakiness
    */
   val workspacePath: os.Path = {
-    Iterator
+    if (sys.env.contains("MILL_TEST_SHARED_OUTPUT_DIR")) baseWorkspacePath
+    else Iterator
       .iterate(1)(_ + 1)
       .map(i => baseWorkspacePath / s"run-$i")
       .find(!os.exists(_))
@@ -39,14 +40,17 @@ trait IntegrationTesterBase {
   def initWorkspace(): Unit = {
     println(s"Copying integration test sources from $workspaceSourcePath to $workspacePath")
     os.makeDir.all(workspacePath)
-    Retry(logger = Retry.printStreamLogger(System.err)) {
-      val tmp = os.temp.dir()
-      val outDir = os.Path(out, workspacePath)
-      if (os.exists(outDir)) os.move.into(outDir, tmp)
-      os.remove.all(tmp)
-    }
+    if (!sys.env.contains("MILL_TEST_SHARED_OUTPUT_DIR"))
+      Retry(logger = Retry.printStreamLogger(System.err)) {
+        val tmp = os.temp.dir()
+        val outDir = os.Path(out, workspacePath)
+        if (os.exists(outDir)) os.move.into(outDir, tmp)
+        os.remove.all(tmp)
+      }
 
-    os.list(workspacePath).foreach(os.remove.all(_))
+    os.list(workspacePath).foreach { p =>
+      if (p.last != "out") os.remove.all(p)
+    }
     val outRelPathOpt = os.FilePath(out) match {
       case relPath: os.RelPath if relPath.ups == 0 => Some(relPath)
       case _ => None
@@ -62,25 +66,27 @@ trait IntegrationTesterBase {
 
     // In case someone manually ran stuff in the integration test workspace earlier,
     // remove any leftover `out/` folder so it does not interfere with the test
-    os.remove.all(workspacePath / "out")
+    if (!sys.env.contains("MILL_TEST_SHARED_OUTPUT_DIR")) os.remove.all(workspacePath / "out")
   }
 
   /**
    * Remove any ID files to try and force them to exit
    */
   def removeProcessIdFile(): Unit = {
-    val outDir = os.Path(out, workspacePath)
-    if (os.exists(outDir)) {
-      if (daemonMode) {
-        val serverPath = outDir / millDaemon
-        os.remove(serverPath / processId)
-      } else {
-        val serverPath0 = outDir / millNoDaemon
+    if (!sys.env.contains("MILL_TEST_SHARED_OUTPUT_DIR")) {
+      val outDir = os.Path(out, workspacePath)
+      if (os.exists(outDir)) {
+        if (daemonMode) {
+          val serverPath = outDir / millDaemon
+          os.remove(serverPath / processId)
+        } else {
+          val serverPath0 = outDir / millNoDaemon
 
-        for (serverPath <- os.list.stream(serverPath0)) os.remove(serverPath / processId)
+          for (serverPath <- os.list.stream(serverPath0)) os.remove(serverPath / processId)
 
+        }
+        Thread.sleep(500) // give a moment for the server to notice the file is gone and exit
       }
-      Thread.sleep(500) // give a moment for the server to notice the file is gone and exit
     }
   }
 }
